maintained init script for avoiding duplicated execution

+++ ch5/init.sh
@@ -1,10 +1,14 @@
 #!/bin/zsh
 
-mkdir $1
-go build ./dispCurrentFolder.go
-var=$(pwd | xargs -I@ ./dispCurrentFolder @)
-cd $1
-cp ~/template/main.go ./
-go mod init $var/$1
-git add main.go go.mod
-git commit -m"initial commit for practice"
+if test ! -e $1 ; then
+    mkdir $1
+    go build ./dispCurrentFolder.go
+    var=$(pwd | xargs -I@ ./dispCurrentFolder @)
+    cd $1
+    cp ~/template/main.go ./
+    go mod init $var/$1
+    git add main.go go.mod
+    git commit -m"initial commit for practice"
+else
+    echo "$1 already exists"
+fi
